// @types/mutation-types.ts
export enum MutationTypes {
  WEB3_SET_IS_CONNECTED = 'SET_IS_CONNECTED',
  WEB3_SET_USER_ADDRESS = 'SET_USER_ADDRESS',
  WEB3_SET_CHAIN_ID = 'SET_CHAIN_ID',
  WEB3_SET_IS_DEFAULT_CHAIN_ID = 'SET_IS_DEFAUL_CHAIN_ID',
  WEB3_SET_NETWORK_NAME = 'SET_NETWORK_NAME',
  LOADER_SET_IS_LOADING = 'SET_IS_LOADING',
  MAIN_SET_THEME = 'SET_THEME'
}